agregamos GestorTarea
